requirements.txt
# 智能用药提醒机器人表情插值生成器依赖包
# 安装命令: pip install -r requirements.txt

# 图像处理库
Pillow>=9.0.0

# 可选：进度条显示（如需要）
# tqdm>=4.64.0

# 可选：更好的字体支持
# fonttools>=4.33.0